Major refactor to allow for parameterized expression
This will enable more generic beats generating prose with
different possible parameters. I accomplished this by inverting the
dependency and creating the shared concept of expressions.

Also simplified concept parameter types because we can use the
required concepts feature for that and remove some complexity.

Also renamed some properties of parameterized concepts to clarify a
bit.

Also stated to introduce the idea of "content packs" that the main.py
code can choose to mix and match to let you choose the tone and
setting of your story, when we have more than just noir.